fix: rename UserRolesCards component and related user roles functionality
