+++ src/views/system/admin/manage-users/partials/UserRolesCards.vue
@@ -2,7 +2,7 @@
 import ConfirmDialog from '@/components/common/ConfirmDialog.vue'
 import UserRolesFormDialog from './UserRolesFormDialog.vue'
 import AppAlert from '@/components/common/AppAlert.vue'
-import { useUserRolesList } from './userRolesList'
+import { useUserRolesCards } from './userRolesCards'
 
 const {
   itemData,
@@ -14,7 +14,7 @@ const {
   onDelete,
   onConfirmDelete,
   userRolesStore,
-} = useUserRolesList()
+} = useUserRolesCards()
 </script>
 
 <template>
@@ -51,7 +51,9 @@ const {
         <v-card-title class="mt-3 d-flex justify-end">
           <v-btn prepend-icon="mdi-tag-plus" color="primary" @click="onAdd"> Add Role </v-btn>
         </v-card-title>
-        <v-card-text class="mt-2 mb-1 text-end"> Add Role if it doesn't exist. </v-card-text>
+        <v-card-text class="mt-2 mb-1 text-end">
+          Create Role if it doesn't already exist.
+        </v-card-text>
       </v-card>
     </v-col>
   </v-row>
